schema_extractor/pyproject.toml
[project]
name = "schema_extractor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyodbc",
    "python-dotenv",
    "langchain[google-genai]",
    "pydantic",
    "langchain-chroma",
    "langchain-community"
]
